Have a general (non-member) compound assignment operator
* clean up old VectorHelper code
* fix Vc_ENABLE_FLOAT_BIT_OPERATORS feature selector

+++ tests/arithmetics.cpp
@@ -213,9 +213,9 @@ TEST_TYPES(Vec, testMulAdd, ALL_TYPES)
     for (std::size_t rep = 0; rep < 10000 / Vec::Size; ++rep) {
         Vec a = Vec::Random();
         if (std::is_floating_point<T>::value) {
-            a *= std::sqrt(std::numeric_limits<T>::max());
+            a *= static_cast<int>(std::sqrt(std::numeric_limits<T>::max()));
         } else if (std::is_signed<T>::value) {
-            a /= std::sqrt(std::numeric_limits<T>::max());
+            a /= static_cast<int>(std::sqrt(std::numeric_limits<T>::max()));
         }
         using ReferenceVector = decltype(a * a);
         ReferenceVector reference = a;
